requirements.txt
google-genai
python-dotenv
gspread
pandas
streamlit>=1.30.0
altair<5
extra-streamlit-components
gallery-dl
Pillow
streamlit-image-coordinates
google-api-python-client
google-auth
